+++ packages/terra-date-picker/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Change the onBlur event to get triggered only when the entire date picker component loses focus.
 
 3.10.0 - (March 26, 2019)
 ------------------