+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Change the onBlur event to get triggered only when the entire date time picker component loses focus.
 
 3.11.0 - (March 26, 2019)
 ------------------